Cleaned up, improved instruction accuracy, bug fixes, and Python fib
Cleaned up a bit, improved the accuracy of the interpreter's instructions relative to mlog, a few bug fixes, and added a Python version of the fibonacci program for the sake of comparison.

+++ src/interpreter.rs
@@ -4,13 +4,14 @@ use std::collections::{BTreeMap, HashMap};
 #[derive(Clone)]
 pub struct VariableTypes {
     pub string: Option<String>,
-    pub float: Option<f32>
+    pub float: Option<f64>
 }
 
 pub struct GlobalVariables<'a> {
     pub position: usize,
     pub variables: VariableMap,
     pub print_buffer: Vec<String>,
+    #[allow(dead_code)]
     pub instruction_map: &'a BTreeMap<String, Callback>
 }
